cron/requirements.txt
datetime == 5.1
psycopg2-binary == 2.9.6
PyYAML == 6.0
ua-parser == 0.16.1
user-agents == 2.2.0